+++ Code/Framework/AzToolsFramework/AzToolsFramework/Entity/EntityUtilityComponent.h
